added studentID insertion API
